+++ model_package/Calibrate/calibration_tools.py
@@ -104,7 +104,7 @@ def isolate_element(quantities, type, elem):
     return(output)
 
 
-def Isbuga_micrormorphic_elasticity_parameters(Emod, nu, Lc):
+def Isbuga_micrormorphic_elasticity_parameters(Emod, nu, Lc, case_1_override=False):
     '''Calculate initial estimate of 18 parameter micromorphic linear elasticity model parameters using method defined in https://doi.org/10.1016/j.ijengsci.2011.04.006
 
     :param float Emod: An estimate of homogenized elastic modulus
@@ -138,6 +138,10 @@ def Isbuga_micrormorphic_elasticity_parameters(Emod, nu, Lc):
     tau_9 = 0.495*(lame_mu*Lc*Lc)
     tau_10 = 0.408*(lame_mu*Lc*Lc)
     tau_11 = 0.495*(lame_mu*Lc*Lc)
+
+    if case_1_override == True:
+        lamb = lame_lambda
+        mu = lame_mu
 
     # collect
     parameters = numpy.array([lamb, mu, eta, tau, kappa, nu_new, sigma,
@@ -527,7 +531,7 @@ def parse_fparams_file(parameter_file, material_type='elastic'):
         params = numpy.hstack([[float(i) for i in UI['line 1'].split(' ')[1:]],
                                [float(i) for i in UI['line 2'].split(' ')[1:]],
                                [float(i) for i in UI['line 3'].split(' ')[1:]],
-                               float(UI['obj'])])
+                               float(UI['obj_func_value'])])
     elif material_type == 'plastic':
         parameter_ordering = plastic_parameter_ordering + elastic_parameter_ordering + ['obj_func_value']
         params = numpy.hstack([[float(i) for i in UI['line 01'].split(' ')[1:]],
@@ -536,7 +540,7 @@ def parse_fparams_file(parameter_file, material_type='elastic'):
                                 [float(i) for i in UI['line 10'].split(' ')[1:]],
                                 [float(i) for i in UI['line 11'].split(' ')[1:]],
                                 [float(i) for i in UI['line 12'].split(' ')[1:]],
-                                float(UI['obj'])])
+                                float(UI['obj_func_value'])])
     elif material_type == 'full_plastic':
         raise NotImplementedError("'full_plastic' option has not been implemented yet!")
     else: